PR2: Fix commit() livelock causing 3m delay on 409 errors (#412)
## Summary
This issue is being addressed in two parts:

* **[PR1](https://github.com/linkedin/openhouse/pull/411):** Updates the
test suite to clearly reproduce and expose the infinite retry loop
(livelock) scenario for documentation and verification purposes.
* **PR2 (this PR):** implement the fix, ensuring the system "fails fast"
by returning a `CommitFailedException` to the client immediately, rather
than trapping the request in a server-side retry loop.

The solution ensures that when commitTransaction() is called, the
existing in-memory metadata is used to build the retry config vs using
the hdfs base metadata (which will always have client retry config).

So, if transaction fails with commitfailedexception, it propagates to
the retryer, where the retries is set to 0, and fails fast back to
client, without changing the client retry behavior and persisted table
retry settings. The in-memory metadata is prevented from being persisted
to the hdfs metadata (and changing client behavior) by removing the
retry property override and replacing it with the original retry config.

This aligns with OpenHouse's design of relying on the House Table
Service (HTS) for optimistic locking and preventing stale writes, and
the client app to resolve retry conflicts.

**Impact**
Kafka ETL can commit faster. 

## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [X] Bug Fixes
- [ ] New Features
- [X] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [X] Tests

For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [ ] Added new tests for the changes made.
- [X] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.