Add multi schema update support for tables during replica table commits (#407)
## Summary

<!--- HINT: Replace #nnn with corresponding Issue number, if you are
fixing an existing issue -->

[Issue](https://github.com/linkedin/openhouse/issues/#nnn) Briefly
discuss the summary of the changes made in this
pull request in 2-3 lines.

When committing table metadatas when multiple schema evolutions have
been performed within a single commit, there can be issues ensuring that
the schema IDs remain consistent. This is more commonly seen in
replication tables as shown below:

```mermaid
---
title: Source Table
---
flowchart TD;
   A[Schema A - id 1] --> B[Schema B - id 2]
   B --> C[Schema C - id 3]
   C --> D[Schema D - id 4]
```

```mermaid
---
title: Replica Table
---
flowchart TD;
   A[Schema A - id 1] --> D[Schema D - id 2]
```
This causes issues because replication uses table snapshots generated
directly on the source table, where each snapshot contains a reference
to its corresponding schema ID. This causes the replica table to be
unreadable for that specific snapshot. This affects certain compute
engines such as Trino, as well as time travel queries.

To resolve this issue, we want to support multiple schema updates within
a single commit on the server side. This resolves an extra field to send
the delta of schemas, identified by schema ID, when performing a
replicated table commit (this can be generalized to not only affect
replica tables). Each schema needs to be serialized in order to ensure
that the column IDs are consistent.

## Changes

- [x] Client-facing API Changes
- [x] Internal API Changes
- [ ] Bug Fixes
- [ ] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [x] Tests


For all the boxes checked, please include additional details of the
changes made in this pull request.

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands
ran, and their output.
- [x] Added new tests for the changes made.
- [x] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please
feel free to ask for help.
- [ ] Some other form of testing like staging or soak time in
production. Please explain.

For all the boxes checked, include a detailed description of the testing
done for the changes made in this pull request.

# Additional Information

- [ ] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the
description.

For all the boxes checked, include additional details of the changes
made in this pull request.